refactor: 숙소 관련 계층형 아키텍처 => 헥사고날 아키텍처 구조 변경
