+++ app/src/main/java/com/hotelJava/accommodation/application/port/DeleteAccommodationQuery.java
@@ -1,0 +1,6 @@
+package com.hotelJava.accommodation.application.port;
+
+public interface DeleteAccommodationQuery {
+
+  void deleteAccommodation(Long accommodationId);
+}
