+++ app/src/main/java/com/hotelJava/accommodation/application/port/SaveAccommodationUseCase.java
@@ -1,0 +1,9 @@
+package com.hotelJava.accommodation.application.port;
+
+import com.hotelJava.accommodation.dto.CreateAccommodationRequestDto;
+import com.hotelJava.accommodation.dto.CreateAccommodationResponseDto;
+
+public interface SaveAccommodationUseCase {
+
+  CreateAccommodationResponseDto saveAccommodation(CreateAccommodationRequestDto command);
+}
