+++ app/src/main/java/com/hotelJava/accommodation/application/port/UpdateAccommodationUseCase.java
@@ -1,0 +1,9 @@
+package com.hotelJava.accommodation.application.port;
+
+import com.hotelJava.accommodation.dto.UpdateAccommodationRequestDto;
+
+public interface UpdateAccommodationUseCase {
+
+  void updateAccommodation(
+      Long accommodationId, UpdateAccommodationRequestDto updateAccommodationRequestDto);
+}
