+++ app/src/main/java/com/hotelJava/accommodation/application/service/AccommodationService.java
@@ -1,23 +1,21 @@
-package com.hotelJava.accommodation.service;
+package com.hotelJava.accommodation.application.service;
 
+import com.hotelJava.accommodation.adapter.persistence.AccommodationRepository;
+import com.hotelJava.accommodation.application.port.DeleteAccommodationQuery;
+import com.hotelJava.accommodation.application.port.SaveAccommodationUseCase;
+import com.hotelJava.accommodation.application.port.UpdateAccommodationUseCase;
 import com.hotelJava.accommodation.domain.Accommodation;
-import com.hotelJava.accommodation.domain.AccommodationType;
 import com.hotelJava.accommodation.dto.CreateAccommodationRequestDto;
 import com.hotelJava.accommodation.dto.CreateAccommodationResponseDto;
-import com.hotelJava.accommodation.dto.FindAccommodationResponseDto;
 import com.hotelJava.accommodation.dto.UpdateAccommodationRequestDto;
-import com.hotelJava.accommodation.picture.domain.Picture;
-import com.hotelJava.accommodation.picture.util.PictureMapper;
-import com.hotelJava.accommodation.repository.AccommodationRepository;
 import com.hotelJava.accommodation.util.AccommodationMapper;
 import com.hotelJava.common.error.ErrorCode;
 import com.hotelJava.common.error.exception.BadRequestException;
-import com.hotelJava.common.error.exception.InternalServerException;
+import com.hotelJava.picture.domain.Picture;
+import com.hotelJava.picture.util.PictureMapper;
 import com.hotelJava.room.domain.Room;
 import com.hotelJava.room.util.RoomMapper;
-import java.time.LocalDate;
 import java.util.List;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -25,7 +23,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class AccommodationService {
+public class AccommodationService implements SaveAccommodationUseCase, UpdateAccommodationUseCase, DeleteAccommodationQuery {
 
   private final AccommodationRepository accommodationRepository;
 
@@ -35,33 +33,8 @@ public class AccommodationService {
 
   private final PictureMapper pictureMapper;
 
-  public List<FindAccommodationResponseDto> findAccommodations(
-      AccommodationType type,
-      String firstLocation,
-      String secondLocation,
-      String name,
-      LocalDate checkInDate,
-      LocalDate checkOutDate,
-      int guestCount) {
-    List<Accommodation> accommodations =
-        accommodationRepository.findAccommodations(
-            type, firstLocation, secondLocation, name, checkInDate, checkOutDate, guestCount);
-    return accommodations.stream()
-        .map(
-            accommodation -> {
-              int minimumRoomPrice =
-                  accommodation.getRooms().stream()
-                      .mapToInt(Room::getPrice)
-                      .min()
-                      .orElseThrow(
-                          () -> new InternalServerException(ErrorCode.NO_MINIMUM_PRICE_FOUND));
-              return accommodationMapper.toFindAccommodationResponseDto(
-                  minimumRoomPrice, accommodation);
-            })
-        .collect(Collectors.toList());
-  }
-
   @Transactional
+  @Override
   public CreateAccommodationResponseDto saveAccommodation(
       CreateAccommodationRequestDto createAccommodationRequestDto) {
     validateDuplicatedAccommodation(createAccommodationRequestDto);
@@ -91,6 +64,7 @@ public class AccommodationService {
   }
 
   @Transactional
+  @Override
   public void updateAccommodation(
       Long accommodationId, UpdateAccommodationRequestDto updateAccommodationRequestDto) {
 
@@ -109,6 +83,7 @@ public class AccommodationService {
   }
 
   @Transactional
+  @Override
   public void deleteAccommodation(Long accommodationId) {
     accommodationRepository
         .findById(accommodationId)
