+++ app/src/main/java/com/hotelJava/picture/domain/Picture.java
@@ -1,4 +1,4 @@
-package com.hotelJava.accommodation.picture.domain;
+package com.hotelJava.picture.domain;
 
 import com.hotelJava.accommodation.domain.Accommodation;
 import com.hotelJava.common.util.BaseTimeEntity;
