+++ app/src/main/java/com/hotelJava/picture/domain/PictureType.java
@@ -1,4 +1,4 @@
-package com.hotelJava.accommodation.picture.domain;
+package com.hotelJava.picture.domain;
 
 public enum PictureType {
   ACCOMMODATION,
