+++ app/src/main/java/com/hotelJava/picture/util/PictureMapper.java
@@ -1,7 +1,7 @@
-package com.hotelJava.accommodation.picture.util;
+package com.hotelJava.picture.util;
 
-import com.hotelJava.accommodation.picture.domain.Picture;
-import com.hotelJava.accommodation.picture.dto.PictureDto;
+import com.hotelJava.picture.domain.Picture;
+import com.hotelJava.picture.dto.PictureDto;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
